optrom.s: Do not use int 14h, use port directly
We store the com port base addr, we find at 40:0 to 40:F0 and remove
it from 40:0 to make DOS unaware of it.

This way we need not use int 14h, nor overwrite it and can directly
use the port for init, read and writs. This might even be faster
than using interrupts.

Signed-off-by: Andreas J. Reichel <andreas@reichel.bayern>